Mark songs which file paths do not exist and skip them
* Create dummy Song object for those songs. These dummy objects only contain the file path (which does not exist), have set the title to the file path (to show the missing path to the user) and a special flag IsFileMissing set on true. The other attributes are set to default or empty values.

* Check if the song to play exists on every call.

* Skip the given song if its file is missing (i.e. the path does not exist).

* Show the missing song path's for the user in red font to make the error visible.

+++ Model/Service/SongImporter.cs
@@ -1,11 +1,29 @@
+using System.IO;
 using Model.DataType;
 
 namespace Model.Service;
 
 public class SongImporter
 {
+    /// <summary>
+    /// Returns a full Song object for the song under the given path.
+    /// <para/>
+    /// If the path does not exist, a dummy object will be returned with the value of <see cref="Song.IsFileMissing"/>
+    /// set to true and with the file path as title so that the error can be shown to the user directly inside a
+    /// playlist.
+    /// </summary>
+    /// <param name="songPath">The song path.</param>
+    /// <returns>The full Song object or a dummy Song object if the file path does not exist.</returns>
     public Song Import(string songPath)
     {
+        // When there is no file under the given path, return a dummy object that indicates that the file is missing.
+        // This dummy object is used because it can still be shown in the View next to the existing songs. So, missing
+        // files will be visible for the user.
+        if (!File.Exists(songPath))
+        {
+            return new Song(songPath);
+        }
+        
         // Read song file
         var songFile = TagLib.File.Create(songPath);
